Invoking Watson via the QAAPI
This code exports the question function, our handler, so that it can be accessed from app.js. The question function takes two parameters, the request, req, and the response, res.
The request, req, is expected to be the string that contains the question text. The question string is then used to create a proper Watson QAAPI question request that conforms to the QAAPI JSON Request and Response model.
Finally, the question function invokes the Watson QAAPI synchronously by using the request module and returns the response "as-is."