Add HSTS to the Connection
The mechanism of HSTS is to send a Strict-Transport-Security header to the client specifying when the SSL policy will expire. The browser will then default to HTTPS when communicating with the application until this header expires.
The header is respected by the browser only if it’s sent over an HTTPS connection and there are no errors with the certificate

Secure Your Node.js Web Application
By: Karl Duuna
Publisher: Pragmatic Bookshelf
Pub. Date: December 28, 2015
Print ISBN-13: 978-1-68050-085-1

Set Up Secure Networking for Node.js Applications