Web Audio API for dynamic sound synthesis
http://www.waudjs.com/
Waud.js - WEB AUDIO LIBRARY WITH HTML5 AUDIO FALLBACK.

https://github.com/rserota/wad

jQuery for your ears http://codecur.io/us/projects/song-demo

The only reason for encoding audio in base64 is to reduce the number of requests which in turn reduces the loading time. It's an alternative to audio sprites. Useful when you have a lot of audio files to load.